Python/requirements.txt
discord-rewrite[voice]
youtube-dl